refactor: 좋아요 기능 수정
- 낙관적 락, Retry 제거 -> native query로 update
- 좋아요 개수 칼럼 int로 수정
- 만약 좋아요 개수가 음수가 되면 동기화 처리

+++ src/main/java/com/tf4/photospot/post/application/response/PostAllResponse.java
@@ -12,7 +12,7 @@ import lombok.Builder;
 public record PostAllResponse(
 	Long id,
 	String detailAddress,
-	Long likeCount,
+	int likeCount,
 	String photoUrl,
 	BubbleResponse bubble,
 	Boolean isLiked,
